refactor(fullaudio): make some of the SIP.js input/output code re-usable
To be used by other bridges
